Move associados features into dedicated app
