+++ accounts/urls.py
@@ -8,18 +8,6 @@ urlpatterns = [
     # Autenticação
     path("login/", views.login_view, name="login"),
     path("logout/", views.logout_view, name="logout"),
-    path("associados/", views.AssociadoListView.as_view(), name="associados_lista"),
-    path(
-        "associados/promover/",
-        views.AssociadoPromoverListView.as_view(),
-        name="associados_promover",
-    ),
-    path(
-        "associados/<int:pk>/promover/form/",
-        views.AssociadoPromoverFormView.as_view(),
-        name="associado_promover_form",
-    ),
-    path("associados/novo/", views.OrganizacaoUserCreateView.as_view(), name="associados_adicionar"),
     # Registro de usuário
     path("resend-confirmation/", views.resend_confirmation, name="resend_confirmation"),
     path("password_reset/", views.password_reset, name="password_reset"),
