+++ associados/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class AssociadosConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "associados"
+    verbose_name = "Associados"
