+++ associados/templates/associados/_promover_card.html
@@ -40,6 +40,6 @@
     </dl>
   </div>
   <footer class="card-footer">
-    <a href="{% url 'accounts:associado_promover_form' user.pk %}" class="btn btn-secondary w-full text-center">{% trans 'Promover' %}</a>
+    <a href="{% url 'associados:associado_promover_form' user.pk %}" class="btn btn-secondary w-full text-center">{% trans 'Promover' %}</a>
   </footer>
 </article>
