+++ associados/templates/associados/hero_action.html
@@ -3,7 +3,7 @@
   {% with tipo=request.user.get_tipo_usuario %}
     {% if tipo == 'admin' or tipo == 'operador' or tipo == 'root' %}
       <div class="flex flex-wrap gap-2">
-        <a href="{% url 'accounts:associados_adicionar' %}?tipo=associado" class="btn btn-secondary">
+        <a href="{% url 'associados:associados_adicionar' %}?tipo=associado" class="btn btn-secondary">
           {% trans 'Adicionar associado' %}
         </a>
       </div>
