+++ tests/accounts/test_associados.py
@@ -29,7 +29,7 @@ def test_admin_list_associados(client):
         is_associado=True,
     )
     client.force_login(admin)
-    resp = client.get(reverse("accounts:associados_lista"))
+    resp = client.get(reverse("associados:associados_lista"))
     assert resp.status_code == 200
     assert assoc.username in resp.content.decode()
 
@@ -39,7 +39,7 @@ def test_search_associados(client):
     create_user("john@example.com", "john", UserType.ASSOCIADO, is_associado=True)
     create_user("jane@example.com", "jane", UserType.ASSOCIADO, is_associado=True)
     client.force_login(admin)
-    resp = client.get(reverse("accounts:associados_lista"), {"q": "john"})
+    resp = client.get(reverse("associados:associados_lista"), {"q": "john"})
     content = resp.content.decode()
     assert "john" in content
     assert "jane" not in content
@@ -61,7 +61,7 @@ def test_coordenador_list_associados(client):
         organizacao=org,
     )
     client.force_login(coord)
-    resp = client.get(reverse("accounts:associados_lista"))
+    resp = client.get(reverse("associados:associados_lista"))
     assert resp.status_code == 200
     assert assoc.username in resp.content.decode()
 
@@ -77,7 +77,7 @@ def test_associados_htmx_returns_partial_grid(client):
         )
 
     client.force_login(admin)
-    url = reverse("accounts:associados_lista")
+    url = reverse("associados:associados_lista")
     resp = client.get(url, HTTP_HX_REQUEST="true")
 
     assert resp.status_code == 200
@@ -133,7 +133,7 @@ def test_associados_filter_actions(client):
     )
 
     client.force_login(admin)
-    url = reverse("accounts:associados_lista")
+    url = reverse("associados:associados_lista")
 
     resp = client.get(url)
     assert resp.status_code == 200
